# docs/swagger.yaml
basePath: /
definitions:
  github_com_DaniaLD_upera-go-test_internal_app_router-handler_product-revision_dto.ProductRevisionsListResponse:
    properties:
      revisions:
        items:
          $ref: '#/definitions/github_com_DaniaLD_upera-go-test_pkg_model.ProductRevision'
        type: array
      total:
        type: integer
    type: object
  github_com_DaniaLD_upera-go-test_internal_app_router-handler_product_dto.CreateProductRequestDto:
    properties:
      color:
        type: string
      description:
        type: string
      imageURL:
        type: string
      name:
        type: string
      price:
        type: integer
    type: object
  github_com_DaniaLD_upera-go-test_internal_app_router-handler_product_dto.ProductCommonResponseDto:
    properties:
      message:
        type: string
      payload:
        $ref: '#/definitions/github_com_DaniaLD_upera-go-test_pkg_model.Product'
      status:
        type: integer
    type: object
  github_com_DaniaLD_upera-go-test_pkg_model.Product:
    properties:
      color:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      imageURL:
        type: string
      name:
        type: string
      price:
        type: integer
      updatedAt:
        type: string
    type: object
  github_com_DaniaLD_upera-go-test_pkg_model.ProductRevision:
    properties:
      createdAt:
        type: string
      id:
        type: string
      newValue:
        $ref: '#/definitions/github_com_DaniaLD_upera-go-test_pkg_model.Product'
      previousValue:
        $ref: '#/definitions/github_com_DaniaLD_upera-go-test_pkg_model.Product'
      productId:
        type: string
      revisionId:
        type: integer
      updatedAttributes:
        items:
          type: string
        type: array
    type: object
info:
  contact:
    email: danial.6274@gmail.com
    name: API Support
  description: This is an assessment for Upera company
  title: Upera - Product Versioning
  version: "1.0"
paths:
  /api/v1/product:
    post:
      description: CreateProduct
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_DaniaLD_upera-go-test_internal_app_router-handler_product_dto.CreateProductRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DaniaLD_upera-go-test_internal_app_router-handler_product_dto.ProductCommonResponseDto'
      summary: CreateProduct
      tags:
      - Product
  /api/v1/product-revision/{productId}:
    get:
      description: GetProductRevisions
      parameters:
      - description: Product Id
        format: string
        in: path
        name: productId
        required: true
        type: string
      - default: 10
        description: limit
        in: query
        name: limit
        type: integer
      - default: 1
        description: page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DaniaLD_upera-go-test_internal_app_router-handler_product-revision_dto.ProductRevisionsListResponse'
      summary: GetProductRevisions
      tags:
      - Product Revision
  /api/v1/product/{id}:
    get:
      description: GetProduct
      parameters:
      - description: Product Id
        format: string
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DaniaLD_upera-go-test_internal_app_router-handler_product_dto.ProductCommonResponseDto'
      summary: GetProduct
      tags:
      - Product
    put:
      description: UpdateProduct
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_DaniaLD_upera-go-test_internal_app_router-handler_product_dto.CreateProductRequestDto'
      - description: Product Id
        format: string
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DaniaLD_upera-go-test_internal_app_router-handler_product_dto.ProductCommonResponseDto'
      summary: UpdateProduct
      tags:
      - Product
  /api/v1/product/{id}/revision/{revisionId}:
    get:
      description: GetProductByRevisionId
      parameters:
      - description: Product Id
        format: string
        in: path
        name: id
        required: true
        type: string
      - description: Revision Id
        format: string
        in: path
        name: revisionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DaniaLD_upera-go-test_internal_app_router-handler_product_dto.ProductCommonResponseDto'
      summary: GetProductByRevisionId
      tags:
      - Product
swagger: "2.0"